requirements.txt
Flask>=3.0.0
elasticsearch>=8.11.0
PyYAML>=6.0.1
python-dateutil>=2.8.2
